+++ src/main/java/fi/dy/masa/malilib/compat/forge/ForgePlatformCompat.java
@@ -4,10 +4,8 @@ import fi.dy.masa.malilib.compat.forge.register.ModConfigScreenRegister;
 
 @Deprecated
 public class ForgePlatformCompat {
-    private static final ThreadLocal<ForgePlatformCompat> forgePlatform = ThreadLocal.withInitial(ForgePlatformCompat::new);
-
     public static ForgePlatformCompat getInstance() {
-        return forgePlatform.get();
+        return new ForgePlatformCompat();
     }
 
     public ModConfigScreenRegister getMod(String id) {