+++ src/main/java/fi/dy/masa/malilib/mixin/compat/MixinForgeGui.java
@@ -1,26 +1,22 @@
-package fi.dy.masa.malilib.mixin;
+package fi.dy.masa.malilib.mixin.compat;
 
-import org.spongepowered.asm.mixin.Final;
+import net.minecraftforge.client.gui.overlay.ForgeGui;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.gui.DrawContext;
-import net.minecraft.client.gui.hud.InGameHud;
 
 import fi.dy.masa.malilib.event.RenderEventHandler;
 
-@Mixin(InGameHud.class)
-public abstract class MixinInGameHud
+@Mixin(ForgeGui.class)
+public abstract class MixinForgeGui
 {
-    @Shadow @Final private MinecraftClient client;
-
     @Inject(method = "render", at = @At("RETURN"))
     private void onGameOverlayPost(DrawContext drawContext, float partialTicks, CallbackInfo ci)
     {
-        ((RenderEventHandler) RenderEventHandler.getInstance()).onRenderGameOverlayPost(drawContext, this.client, partialTicks);
+        ((RenderEventHandler) RenderEventHandler.getInstance()).onRenderGameOverlayPost(drawContext, MinecraftClient.getInstance(), partialTicks);
     }
 }